adjusted formatting while familiarizing myself with issue #1's resolution and how it iterracts with possible solutions for issue #3
