+++ src/App.css
@@ -1,4 +1,6 @@
 .App {
+  font-size: 62.5%;
+  color: black;
   text-align: center;
 }
 
@@ -13,8 +15,9 @@
   flex-direction: column;
   align-items: center;
   justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
+  font-size: 1.6rem;
+  /* font-size: calc(10px + 2vmin); */
+  color: black;
 }
 
 .App-link {
@@ -25,6 +28,7 @@
 }
 
 nav {
+  font-size: 1.6rem;
   position: absolute;
   width: 100%;
   bottom: 0;
