+++ src/components/List.js
@@ -1,7 +1,25 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
+import { db } from '../lib/firebase';
 
 const List = () => {
-  return <h1>THIS IS THE LIST</h1>;
+  let [results, setResults] = useState([]);
+  useEffect(() => {
+    db.collection('shopping-list').onSnapshot(function(querySnapshot) {
+      let querySnapshotResults = [];
+      querySnapshot.forEach(function(doc) {
+        const { id, name } = doc.data();
+        querySnapshotResults.push({ id, name });
+      });
+      setResults(querySnapshotResults);
+    });
+  }, []);
+  return (
+    <ul style={{ color: 'black' }}>
+      {results.map(result => (
+        <li key={result.id}>{result.name}</li>
+      ))}
+    </ul>
+  );
 };
 
 export default List;
