+++ src/components/Welcome.js
@@ -1,0 +1,11 @@
+import React, { useState, useEffect } from 'react';
+
+const Welcome = () => {
+  return (
+    <div>
+      <h1>Welcome to your Smart Shopping List</h1>
+    </div>
+  );
+};
+
+export default Welcome;
